🐞 fix(guided_diffusion/tain_util.py): MPI resume freeze
When using MPI to start multi-card training and resume from a checkpoint, the training program will hang. The reason is that the original code uses MPI broadcast parameters in the main process, but the child process uses torch.dist to receive parameters. Fixed so that all processes communicate using MPI.